# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/profiles":{"put":{"tags":["用户资料管理"],"summary":"更新当前用户信息","operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultProfileVO"}}}}}}},"/articles/{id}":{"get":{"tags":["文章/评论管理"],"summary":"获取文章详情","operationId":"getArticle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultArticleVO"}}}}}},"put":{"tags":["文章/评论管理"],"summary":"更新文章","operationId":"updateArticle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultArticleVO"}}}}}},"delete":{"tags":["文章/评论管理"],"summary":"删除文章","operationId":"deleteArticle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/users/register":{"post":{"tags":["用户管理"],"summary":"用户注册","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegisterDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultUserLoginVO"}}}}}}},"/users/login":{"post":{"tags":["用户管理"],"summary":"用户登录","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultUserLoginVO"}}}}}}},"/profiles/follow/{id}":{"post":{"tags":["用户资料管理"],"summary":"关注用户","operationId":"follow","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/articles":{"post":{"tags":["文章/评论管理"],"summary":"创建文章","operationId":"createArticle","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultInteger"}}}}}}},"/articles/favorite/{id}":{"post":{"tags":["文章/评论管理"],"summary":"收藏文章","operationId":"favoriteArticle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/articles/comments/{id}":{"post":{"tags":["文章/评论管理"],"summary":"添加评论","operationId":"addComment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/tags":{"get":{"tags":["标签管理"],"summary":"获取标签列表","operationId":"getTags","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/profiles/{id}":{"get":{"tags":["用户资料管理"],"summary":"获取用户资料","operationId":"getProfile","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultProfileVO"}}}}}}},"/articles/list":{"get":{"tags":["文章/评论管理"],"summary":"获取文章卡片列表","operationId":"getArticles","parameters":[{"name":"authorId","in":"query","description":"文章作者id","required":false,"schema":{"type":"string","description":"文章作者id"}},{"name":"tagList","in":"query","description":"文章标签","required":false,"schema":{"type":"string","description":"文章标签"}},{"name":"limit","in":"query","description":"页码","required":false,"schema":{"type":"string","description":"页码","default":"1"}},{"name":"offset","in":"query","description":"每页记录数","required":false,"schema":{"type":"string","description":"每页记录数","default":"5"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResultArticleCardVO"}}}}}}},"/articles/feed":{"get":{"tags":["文章/评论管理"],"summary":"获取文章卡片列表（登录校验）","operationId":"getArticlesFees","parameters":[{"name":"authorId","in":"query","description":"文章作者id","required":false,"schema":{"type":"string","description":"文章作者id"}},{"name":"tagList","in":"query","description":"文章标签","required":false,"schema":{"type":"string","description":"文章标签"}},{"name":"limit","in":"query","description":"页码","required":false,"schema":{"type":"string","description":"页码","default":"1"}},{"name":"offset","in":"query","description":"每页记录数","required":false,"schema":{"type":"string","description":"每页记录数","default":"5"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResultArticleCardVO"}}}}}}},"/articles/comments":{"get":{"tags":["文章/评论管理"],"summary":"获取评论","operationId":"getComments","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListCommentVO"}}}}}}},"/articles/comments/{commentId}":{"delete":{"tags":["文章/评论管理"],"summary":"删除评论","operationId":"deleteComment","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}}},"components":{"schemas":{"UserUpdateDTO":{"required":["username"],"type":"object","properties":{"username":{"type":"string","description":"用户名"},"bio":{"type":"string","description":"个人简介"},"avatar":{"type":"string","description":"头像url"}}},"ProfileVO":{"type":"object","properties":{"id":{"type":"integer","description":"用户ID","format":"int32"},"username":{"type":"string","description":"用户名"},"bio":{"type":"string","description":"用户简介"},"avatar":{"type":"string","description":"用户头像链接"},"fansCount":{"type":"integer","description":"粉丝数量","format":"int32"},"followCount":{"type":"integer","description":"关注数量","format":"int32"},"following":{"type":"boolean","description":"是否已关注该用户"}}},"ResultProfileVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/ProfileVO"}}},"ArticleUpdateDTO":{"required":["body","description","title"],"type":"object","properties":{"title":{"type":"string","description":"文章标题"},"description":{"type":"string","description":"文章描述"},"body":{"type":"string","description":"文章内容"}}},"ArticleVO":{"type":"object","properties":{"id":{"type":"integer","description":"文章ID","format":"int32"},"title":{"type":"string","description":"文章标题"},"description":{"type":"string","description":"文章描述"},"body":{"type":"string","description":"文章正文内容"},"tags":{"type":"string","description":"文章标签列表"},"createdAt":{"type":"string","description":"文章创建时间"},"favorited":{"type":"boolean","description":"是否已收藏"},"liked":{"type":"boolean","description":"是否已点赞"},"author":{"$ref":"#/components/schemas/ProfileVO"}}},"ResultArticleVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/ArticleVO"}}},"UserRegisterDTO":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string","description":"用户名"},"password":{"type":"string","description":"密码"}}},"ResultUserLoginVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/UserLoginVO"}}},"UserLoginVO":{"type":"object","properties":{"id":{"type":"integer","description":"用户ID","format":"int32"},"username":{"type":"string","description":"用户名"},"bio":{"type":"string","description":"用户简介"},"avatar":{"type":"string","description":"用户头像链接"},"followCount":{"type":"integer","description":"关注数量","format":"int32"},"fansCount":{"type":"integer","description":"粉丝数量","format":"int32"},"token":{"type":"string","description":"用户登录令牌"}}},"UserLoginDTO":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string","description":"用户名"},"password":{"type":"string","description":"密码"}}},"ResultVoid":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"object"}}},"ArticleCreateDTO":{"required":["body","description","title"],"type":"object","properties":{"title":{"type":"string","description":"文章标题"},"description":{"type":"string","description":"文章描述"},"body":{"type":"string","description":"文章内容"},"tagList":{"type":"array","description":"文章标签","items":{"type":"string","description":"文章标签"}}}},"ResultInteger":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"integer","format":"int32"}}},"CommentDTO":{"required":["body"],"type":"object","properties":{"body":{"type":"string","description":"评论内容"}}},"ResultString":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"string"}}},"ArticleCardVO":{"type":"object","properties":{"id":{"type":"integer","description":"文章ID","format":"int32"},"title":{"type":"string","description":"文章标题"},"description":{"type":"string","description":"文章描述"},"body":{"type":"string","description":"文章正文内容"},"likeCount":{"type":"integer","description":"点赞数量","format":"int32"},"favoritesCount":{"type":"integer","description":"收藏数量","format":"int32"},"createdAt":{"type":"string","description":"文章创建时间"},"tags":{"type":"string","description":"标签列表"},"author":{"type":"string","description":"作者用户名"},"avatar":{"type":"string","description":"作者头像链接"},"liked":{"type":"boolean","description":"是否已点赞"},"favorited":{"type":"boolean","description":"是否已收藏"}}},"PageResultArticleCardVO":{"type":"object","properties":{"total":{"type":"integer","format":"int64"},"records":{"type":"array","items":{"$ref":"#/components/schemas/ArticleCardVO"}},"page":{"type":"integer","format":"int64"},"pageSize":{"type":"integer","format":"int64"}}},"CommentVO":{"type":"object","properties":{"id":{"type":"integer","description":"评论ID","format":"int32"},"body":{"type":"string","description":"评论内容"},"createdAt":{"type":"string","description":"评论创建时间"},"username":{"type":"string","description":"评论用户名称"},"avatar":{"type":"string","description":"评论用户头像链接"}}},"ResultListCommentVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/CommentVO"}}}}}}}